Builder Welcome now shows config preview card

+++ src/lib/dragon/DragonConfigPreview.svelte
@@ -8,6 +8,7 @@
 	const dispatch = createEventDispatcher<{ clickDelete: DragonConfig }>();
 
 	export let config: DragonConfig;
+	export let showDeleteButton: boolean = true;
 </script>
 
 <div class="daisy-card daisy-card-compact w-full bg-white">
@@ -18,14 +19,16 @@
 		<h2 class="daisy-card-title">{config.getTitle()}</h2>
 		<p>{config.toString()}</p>
 		<div class="daisy-card-actions justify-end">
-			<button
-				class="daisy-btn daisy-btn-outline hover:daisy-btn-error"
-				on:click={() => {
-					dispatch('clickDelete', config);
-				}}
-			>
-				Delete Dragon
-			</button>
+			{#if showDeleteButton}
+				<button
+					class="daisy-btn daisy-btn-outline hover:daisy-btn-error"
+					on:click={() => {
+						dispatch('clickDelete', config);
+					}}
+				>
+					Delete Dragon
+				</button>
+			{/if}
 			<button
 				class="daisy-btn daisy-btn-neutral text-white"
 				on:click={() => {
